ACM_ICPC problem: LargestSum
Description
Given a triangle of positive integers, we want to find a largest possible sum along a path from top
row to bottom row. A path can go down either directly south or directly south-east (Note, you can
NEVER go EAST, WEST, or NORTH etc). For example, in the triangle
1
2 3
1 5 7
9 1 1 2
8 5 4 13 2
the path 1→3→7→2→13 is a legal path with sum 26. This path is “best” because 26 is the largest
such sum. Generally, suppose that the triangle is stored in an array Ann [1.. ,1.. ]with the relevant
entries being on the diagonal and below. For 1≤ ≤r n and 1≤ ≤c r let Brc [, ]be the sum along the
best path that ends in row r and column c of the matrix A.
Example: B[1,1] = 1, B[2,1] = 3, B[2,2] = 4
2 Input
There are several sets of triangle data.
For each set of triangle data, the first number is the number of rows (n) in the triangle. This is
followed by n rows of data. Row i has i cols of data. All data is integer type.
3 Output
There are two lines of data for EACH set of triangle data. The first line states "Max is ???" where
??? is the largest path sum. The second line shows the path. (See the Sample).
4 Samples
INPUT
2
18
430 302
6
102
76 76
268 153 185
492 480 112 405
286 436 126 25 391
390 32 399 76 29 174
OUTPUT
Max is 448
18-->430
Max is 1773
102-->76-->268-->492-->436-->399